% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{rotterdam}
\alias{rotterdam}
\title{Rotterdam dataset from survival}
\format{
A data frame with 2982 rows and 15 variables
See \link[survival]{rotterdam}
}
\usage{
rotterdam
}
\description{
Rotterdam dataset from survival
}
\keyword{datasets}
