% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drkm.R
\name{sjovan_survdiff}
\alias{sjovan_survdiff}
\title{Estimate the survival curves in two treatment groups}
\usage{
sjovan_survdiff(
  oformula = NULL,
  ofunc = "coxph",
  oarg = list(),
  cformula = NULL,
  cfunc = "coxph",
  carg = list(),
  eformula = NULL,
  earg = list(),
  method = "DR",
  times = NULL,
  rel.tol = .Machine$double.eps^0.1,
  jacobian.method = "simple",
  se.type = "none",
  data = NULL,
  weights = NULL,
  subset = NULL,
  R = 50,
  parallel = "no",
  cl = NULL,
  ncpus = NULL
)
}
\arguments{
\item{oformula}{The outcome formula, the left side should be a call to \link[survival]{Surv}}

\item{ofunc}{The model type for the outcome, one of "coxph" or "survreg"}

\item{oarg}{Arguments passed to \code{ofunc}}

\item{cformula}{The censoring model formula, the left side should be a call to \link[survival]{Surv}}

\item{cfunc}{The model type for the censoring model, either "coxph" or "survreg"}

\item{carg}{Arguments passed to \code{cfunc}}

\item{eformula}{The exposure model formula, which will be fit by logistic regression in a call to \link[stats]{glm}}

\item{earg}{Arguments passed to the glm for the exposure model}

\item{method}{Estimation method, either "IPW" or "DR"}

\item{times}{Vector of times at which to estimate the survival curves}

\item{rel.tol}{Convergence tolerance}

\item{jacobian.method}{Method of computing the jacobian, passed to \link[numDeriv]{jacobian}}

\item{se.type}{Method of calculating standard errors, either "none" for no standard errors, "sandwich", or "boot"}

\item{data}{Data frame in which to find the variables in the model formulas}

\item{weights}{Vector of case weights}

\item{subset}{Logical vector}

\item{R}{Number of bootstrap replicates, used if \code{se.type = "boot"}}

\item{parallel}{Parallel processing for bootstrap, see \link[boot]{boot}}

\item{cl}{Optional cluster if parallel processing, see \link[boot]{boot}}

\item{ncpus}{Number of cpus to use for parallel processing, see \link[boot]{boot}}
}
\value{
A list with the estimates and standard errors, if requested
}
\description{
Doubly robust estimation of the survival curves in two exposure groups
and their difference using the method of Sjolander and Vansteelandt.
}
\examples{
df <- rotterdam
df$time <- pmin(df$rtime, df$dtime) / 365.25
df$status <- ifelse(df$recur == 1 | df$death == 1, 1, 0)
df$censor <- 1 - df$status
drFit <-
  sjovan_survdiff(
    oformula = Surv(time, status) ~ chemo + year + age + meno +
      size + factor(grade) + nodes + pgr + er + hormon,
    ofunc = "survreg",
    cformula = Surv(time, censor) ~ chemo + year + age,
    cfunc = "survreg",
    eformula = chemo ~ year + age + meno + size +
      factor(grade) + nodes + pgr + er + hormon,
    method = "DR",
    times = c(2.5, 5, 7.5),
    se.type = "sandwich",
    data = df
  )
drFit

}
