% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drkm.R
\name{sjovan_survdiff}
\alias{sjovan_survdiff}
\title{Estimate the survival curves in two treatment groups}
\usage{
sjovan_survdiff(
  oformula = NULL,
  ofunc = "coxph",
  oarg = list(),
  cformula = NULL,
  cfunc = "coxph",
  carg = list(),
  eformula = NULL,
  earg = list(),
  method = "DR",
  times = NULL,
  rel.tol = .Machine$double.eps^0.1,
  jacobian.method = "simple",
  se.type = "none",
  data = NULL,
  weights = NULL,
  subset = NULL,
  R = 50,
  parallel = "no",
  cl = NULL,
  ncpus = NULL
)
}
\arguments{
\item{oformula}{The outcome formula, the left side should be a call to \link[survival]{Surv}}

\item{ofunc}{The model type for the outcome, one of "coxph" or "survreg"}

\item{oarg}{Arguments passed to \code{ofunc}}

\item{cformula}{The censoring model formula, the left side should be a call to \link[survival]{Surv}}

\item{cfunc}{The model type for the censoring model, either "coxph" or "survreg"}

\item{carg}{Arguments passed to \code{cfunc}}

\item{eformula}{The exposure model formula, which will be fit by logistic regression in a call to \link[stats]{glm}}

\item{earg}{Arguments passed to the glm for the exposure model}

\item{method}{Estimation method, either "IPW" or "DR"}

\item{times}{Vector of times at which to estimate the survival curves}

\item{rel.tol}{Convergence tolerance}

\item{jacobian.method}{Method of computing the jacobian, passed to \link[numDeriv]{jacobian}}

\item{se.type}{Method of calculating standard errors, either "none" for no standard errors, "sandwich", or "boot"}

\item{data}{Data frame in which to find the variables in the model formulas}

\item{weights}{Vector of case weights}

\item{subset}{Logical vector}

\item{R}{Number of bootstrap replicates, used if \code{se.type = "boot"}}

\item{parallel}{Parallel processing for bootstrap, see \link[boot]{boot}}

\item{cl}{Optional cluster if parallel processing, see \link[boot]{boot}}

\item{ncpus}{Number of cpus to use for parallel processing, see \link[boot]{boot}}
}
\value{
A list with the estimates and standard errors, if requested
}
\description{
Doubly robust estimation of the survival curves in two exposure groups
and their difference using the method of Sjolander and Vansteelandt (2017) .
}
\details{
Following Bai et al. 2013 and Sjölander and Vansteelandt 2017, consider the
following estimating equation for \eqn{S_x(t) = p\{T(x) > t\}}:
\deqn{\sum_{i = 1}^n \left[S_x(t) - \frac{I_{X_i = x}
I_{\tilde{T}_i > t}}{\bar{g}(\boldsymbol{Z}_i, \boldsymbol{\alpha})G_c(t,
X_i, \boldsymbol{Z}_i)} - \frac{I_{X_i = x} - \bar{g}(\boldsymbol{Z}_i,
\boldsymbol{\alpha})}{\bar{g}(\boldsymbol{Z}_i, \boldsymbol{\alpha})} H(t,
\boldsymbol{Z}_i, X = x) - \right. \\ \left. \frac{I_{X_i = x} -
\bar{g}(\boldsymbol{Z}_i, \boldsymbol{\alpha})}{\bar{g}(\boldsymbol{Z}_i,
\boldsymbol{\alpha})} H(t, \boldsymbol{Z}_i, X = x) \int_0^t\frac{d\, M_c(u,
\boldsymbol{Z}_i, X_i, \tilde{T}_i, \Delta_i)}{G_c(u, X_i,
\boldsymbol{Z}_i)H(u, \boldsymbol{Z}_i, X_i)}\right] = 0, }
where \eqn{\bar{g}(\boldsymbol{Z}_i, \boldsymbol{\alpha}) = g(\boldsymbol{Z}_i,
\boldsymbol{\alpha})^x (1 - g(\boldsymbol{Z}_i,
\boldsymbol{\alpha}))^{(1-x)}}, \eqn{H(t, \boldsymbol{Z}, X)} is a model for
\eqn{p\{T > t | \boldsymbol{Z}, X\}}, and \eqn{M_c(t, \boldsymbol{Z}, X, \tilde{T},
\Delta)} is the martingale increment for the censoring distribution. The
above is an unbiased estimating equation for \eqn{S_x(t)} if either \eqn{H(t,
\boldsymbol{Z}, X)} is correctly specified for both censoring and
confounding, i.e., is a correctly specified model for \eqn{p\{T(x) > t |
\boldsymbol{Z}, X\}} or both \eqn{G_c(u, X, \boldsymbol{Z})} and
\eqn{g(\boldsymbol{Z}, \boldsymbol{\alpha})} are correctly specified for
censoring and confounding, respectively. To obtain estimates of the
difference in survival probabilities, one must specify models for the unknown
functions \eqn{g}, \eqn{G_c}, and \eqn{H}, get estimates of those, plug them
into the estimating equations, and solve for \eqn{S_x(t)} under \eqn{x \in
\{0, 1\}}. In this package, one can use semiparametric Cox models or
parametric survival models for the outcome and the censoring distributions,
and logistic regression for the propensity score model. Bai et al. 2013
provide an expression for a variance estimator that accounts for the
uncertainty due to the estimation of the propensity score \eqn{g} and the
censoring distribution \eqn{G_c}.
}
\examples{
df <- rotterdam
df$time <- pmin(df$rtime, df$dtime) / 365.25
df$status <- ifelse(df$recur == 1 | df$death == 1, 1, 0)
df$censor <- 1 - df$status
drFit <-
  sjovan_survdiff(
    oformula = Surv(time, status) ~ chemo + year + age + meno +
      size + factor(grade) + nodes + pgr + er + hormon,
    ofunc = "survreg",
    cformula = Surv(time, censor) ~ chemo + year + age,
    cfunc = "survreg",
    eformula = chemo ~ year + age + meno + size +
      factor(grade) + nodes + pgr + er + hormon,
    method = "DR",
    times = c(2.5, 5, 7.5),
    se.type = "sandwich",
    data = df
  )
drFit

}
\references{
Xiaofei Bai, Anastasios A Tsiatis, and Sean~M O'Brien.
  "Doubly-robust estimators of treatment-specific survival distributions in
  observational studies with stratified sampling." Biometrics,
  69(4):830--839, 2013.

  Sjölander, Arvid, and Stijn Vansteelandt. "Doubly
  robust estimation of attributable fractions in survival analysis."
  Statistical methods in medical research 26(2):948-969, 2017.
}
